Вынос общей логики символов переменных в абстрактный класс
